+++ client/plasmacli/eth/balance.go
@@ -2,8 +2,7 @@ package eth
 
 import (
 	"fmt"
-
-	ks "github.com/FourthState/plasma-mvp-sidechain/client/keystore"
+	"github.com/FourthState/plasma-mvp-sidechain/client/store"
 	ethcmn "github.com/ethereum/go-ethereum/common"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -12,20 +11,25 @@ import (
 func init() {
 	queryCmd.AddCommand(balanceCmd)
 	balanceCmd.Flags().StringP(addrF, "A", "", "query based on address")
-	viper.BindPFlags(balanceCmd.Flags())
 }
 
 var balanceCmd = &cobra.Command{
 	Use:   "balance <account>",
 	Short: "Query for balance avaliable for withdraw from rootchain",
-	Args:  cobra.MaximumNArgs(1),
+	Long: `Query for balance avaliable for withdrawal from rootchain.
+	
+Usage: 
+	plasmacli eth query balance <account>
+	plasmacli eth query balance --address <address>`,
+	Args: cobra.MaximumNArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) (err error) {
+		viper.BindPFlags(cmd.Flags())
 		var addr ethcmn.Address
 
 		if viper.GetString(addrF) != "" {
 			addr = ethcmn.HexToAddress(viper.GetString(addrF))
 		} else if len(args) > 0 {
-			if addr, err = ks.Get(args[0]); err != nil {
+			if addr, err = store.GetAccount(args[0]); err != nil {
 				return fmt.Errorf("failed to retrieve account: { %s }", err)
 			}
 		} else {