+++ client/plasmacli/eth/block.go
@@ -2,18 +2,16 @@ package eth
 
 import (
 	"fmt"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 	"math/big"
 	"strconv"
 	"time"
-
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
 func init() {
 	queryCmd.AddCommand(blockCmd)
 	blockCmd.Flags().String(limitF, "1", "number of plasma blocks to be displayed")
-	viper.BindPFlags(blockCmd.Flags())
 }
 
 var blockCmd = &cobra.Command{
@@ -27,6 +25,7 @@ Usage:
 	plasmacli eth query block <number> --limit <number>`,
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
+		viper.BindPFlags(cmd.Flags())
 		curr, err := strconv.ParseInt(args[0], 10, 64)
 		if err != nil {
 			return fmt.Errorf("failed to parse block number: { %s }", err)
@@ -39,14 +38,19 @@ Usage:
 
 		end := curr + lim
 		for curr < end {
-			block, err := rc.session.PlasmaChain(big.NewInt(curr))
+			block, err := rc.contract.PlasmaChain(nil, big.NewInt(curr))
 			if err != nil {
 				return fmt.Errorf("failed to retrieve block: { %s }", err)
 			}
+			if block.CreatedAt.Int64() == 0 {
+				break
+			}
+
+			fmt.Printf("Block: %d\nHeader: 0x%x\nTxs: %d\nFee: %d\nCreated: %v\n\n",
+				curr, block.Header, block.NumTxns, block.FeeAmount, time.Unix(block.CreatedAt.Int64(), 0))
 			curr++
-			fmt.Printf("Header: 0x%x\nTxs: %d\nFee: %d\nCreated: %v\n",
-				block.Header, block.NumTxns, block.FeeAmount, time.Unix(block.CreatedAt.Int64(), 0))
 		}
+
 		return nil
 	},
 }