+++ client/plasmacli/eth/deposit.go
@@ -2,29 +2,32 @@ package eth
 
 import (
 	"fmt"
-	"math/big"
-	"strconv"
-
-	ks "github.com/FourthState/plasma-mvp-sidechain/client/keystore"
+	"github.com/FourthState/plasma-mvp-sidechain/client/store"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"math/big"
+	"strconv"
 )
 
 func init() {
 	ethCmd.AddCommand(depositCmd)
-	depositCmd.Flags().StringP(gasLimitF, "g", "21000", "gas limit for ethereum transaction")
-	viper.BindPFlags(depositCmd.Flags())
+	depositCmd.Flags().StringP(gasLimitF, "g", "150000", "gas limit for ethereum transaction")
 }
 
 var depositCmd = &cobra.Command{
 	Use:   "deposit <amount> <account>",
 	Short: "Deposit to rootchain contract",
-	Long:  `Deposit to the rootchain contract as specified in plasma.toml.`,
-	Args:  cobra.ExactArgs(2),
+	Long: `Deposit to the rootchain contract as specified in plasma.toml.
+
+Usage:
+	plasmacli eth deposit <amount> <account> --gas-limit 30000`,
+	Args: cobra.ExactArgs(2),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		key, err := ks.GetKey(args[1])
+		viper.BindPFlags(cmd.Flags())
+
+		key, err := store.GetKey(args[1])
 		if err != nil {
 			return fmt.Errorf("failed to retrieve account: { %s }", err)
 		}
@@ -39,24 +42,21 @@ var depositCmd = &cobra.Command{
 			return fmt.Errorf("failed to parse gas limit: { %s }", err)
 		}
 
-		// bind key to session
-		// revert after use
+		// bind key, generate transact opts
 		auth := bind.NewKeyedTransactor(key)
-		defer func() {
-			rc.session.TransactOpts = bind.TransactOpts{}
-		}()
-		rc.session.TransactOpts = bind.TransactOpts{
+		transactOpts := &bind.TransactOpts{
 			From:     auth.From,
 			Signer:   auth.Signer,
 			GasLimit: gasLimit,
 			Value:    big.NewInt(amt),
 		}
 
-		if _, err := rc.session.Deposit(crypto.PubkeyToAddress(key.PublicKey)); err != nil {
+		tx, err := rc.contract.Deposit(transactOpts, crypto.PubkeyToAddress(key.PublicKey))
+		if err != nil {
 			return fmt.Errorf("failed to deposit: { %s }", err)
 		}
 
-		fmt.Printf("Successfully deposited\n")
+		fmt.Printf("Successfully sent deposit transaction\nTransaction Hash: 0x%x\n", tx.Hash())
 		return nil
 	},
 }