+++ client/plasmacli/eth/main.go
@@ -2,18 +2,16 @@ package eth
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
-
 	config "github.com/FourthState/plasma-mvp-sidechain/client"
-	ks "github.com/FourthState/plasma-mvp-sidechain/client/keystore"
+	"github.com/FourthState/plasma-mvp-sidechain/client/store"
 	contracts "github.com/FourthState/plasma-mvp-sidechain/contracts/wrappers"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	ethcmn "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"os"
+	"path/filepath"
 )
 
 const (
@@ -26,6 +24,7 @@ const (
 	gasLimitF  = "gas-limit"
 	indexF     = "index"
 	limitF     = "limit"
+	ownerF     = "owner"
 	positionF  = "position"
 	proofF     = "proof"
 	sigsF      = "signatures"
@@ -33,12 +32,12 @@ const (
 	txBytesF   = "tx-bytes"
 
 	minExitBond = 200000 // specified by rootchain contract
+	oneWeek     = 168    // 168 hours in a week
 )
 
 type Plasma struct {
 	ec *ethclient.Client
 
-	session  *contracts.PlasmaMVPSession
 	contract *contracts.PlasmaMVP
 
 	nodeURL string // current url set for the eth client
@@ -62,7 +61,7 @@ func EthCmd() *cobra.Command {
 // Update ethereum client connection if params have changed
 func persistentPreRunEFn() func(*cobra.Command, []string) error {
 	return func(cmd *cobra.Command, args []string) error {
-		plasmaConfigFilePath := filepath.Join(viper.GetString(ks.DirFlag), "plasma.toml")
+		plasmaConfigFilePath := filepath.Join(viper.GetString(store.DirFlag), "plasma.toml")
 
 		if _, err := os.Stat(plasmaConfigFilePath); os.IsNotExist(err) {
 			plasmaConfig := config.DefaultPlasmaConfig()
@@ -81,7 +80,7 @@ func persistentPreRunEFn() func(*cobra.Command, []string) error {
 
 		// Check to see if the eth connection params have changed
 		if conf.EthNodeURL == "" {
-			return fmt.Errorf("please specify a node url for eth connection in %s/plasma.toml", viper.GetString(ks.DirFlag))
+			return fmt.Errorf("please specify a node url for eth connection in %s/plasma.toml", viper.GetString(store.DirFlag))
 		} else if rc.nodeURL != conf.EthNodeURL {
 			if err := initEthConn(conf); err != nil {
 				return err
@@ -101,22 +100,14 @@ func initEthConn(conf config.PlasmaConfig) error {
 	}
 	ec := ethclient.NewClient(c)
 
-	// Create a session with the contract and operator account
+	// Bind rootchain contract and operator account
 	plasmaContract, err := contracts.NewPlasmaMVP(ethcmn.HexToAddress(conf.EthPlasmaContractAddr), ec)
 	if err != nil {
 		return fmt.Errorf("failed to bind to contract: { %s }", err)
 	}
 
-	session := &contracts.PlasmaMVPSession{
-		Contract: plasmaContract,
-		CallOpts: bind.CallOpts{
-			Pending: true,
-		},
-	}
-
 	rc = Plasma{
 		ec:       ec,
-		session:  session,
 		contract: plasmaContract,
 		nodeURL:  conf.EthNodeURL,
 	}