+++ client/plasmacli/keys/import.go
@@ -4,19 +4,20 @@ import (
 	"crypto/ecdsa"
 	"errors"
 	"fmt"
-	ks "github.com/FourthState/plasma-mvp-sidechain/client/keystore"
+	"github.com/FourthState/plasma-mvp-sidechain/client/store"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
+// Flags
 const (
-	flagFile = "file"
+	fileF = "file"
 )
 
 func init() {
 	keysCmd.AddCommand(importCmd)
-	importCmd.Flags().String(flagFile, "", "read the private key from the specified keyfile (must be absolute path)")
+	importCmd.Flags().String(fileF, "", "read the private key from the specified keyfile (must be absolute path)")
 	viper.BindPFlags(importCmd.Flags())
 }
 
@@ -43,30 +44,30 @@ You must remember this passphrase to unlock your account in the future.
 
 		var key *ecdsa.PrivateKey
 		var err error
-		file := viper.GetString(flagFile)
+		file := viper.GetString(fileF)
 		if file != "" {
 			key, err = crypto.LoadECDSA(file)
 			if err != nil {
-				return fmt.Errorf("failed loading the keyfile : %s", err)
+				return fmt.Errorf("failed loading the keyfile: { %s }", err)
 			}
 		} else {
 			if len(args) < 2 {
-				return errors.New("Please provide an unencrytped private if the --file flag is not set")
+				return errors.New("please provide an unencrytped private if the --file flag is not set")
 			}
 			key, err = crypto.HexToECDSA(args[1])
 			if err != nil {
-				return fmt.Errorf("failed parsing private key: %s", err)
+				return fmt.Errorf("failed parsing private key: { %s }", err)
 			}
 
 		}
 
-		address, err := ks.ImportECDSA(name, key)
+		address, err := store.ImportECDSA(name, key)
 		if err != nil {
 			return err
 		}
 
 		fmt.Println("Successfully imported.")
-		fmt.Printf("NAME: %s\t\tADDRESS: %v\n", name, address.Hex())
+		fmt.Printf("NAME: %s\t\tADDRESS: 0x%x\n", name, address)
 		return nil
 	},
 }