+++ client/plasmacli/keys/update.go
@@ -1,18 +1,20 @@
 package keys
 
 import (
-	ks "github.com/FourthState/plasma-mvp-sidechain/client/keystore"
+	"fmt"
+	"github.com/FourthState/plasma-mvp-sidechain/client/store"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
+// Flags
 const (
-	flagName = "name"
+	nameF = "name"
 )
 
 func init() {
 	keysCmd.AddCommand(updateCmd)
-	updateCmd.Flags().String(flagName, "", "updated key name.")
+	updateCmd.Flags().String(nameF, "", "updated key name.")
 	viper.BindPFlags(updateCmd.Flags())
 }
 
@@ -29,11 +31,13 @@ Usage:
 	RunE: func(cmd *cobra.Command, args []string) error {
 		name := args[0]
 
-		updatedName := viper.GetString(flagName)
-		if err := ks.Update(name, updatedName); err != nil {
+		updatedName := viper.GetString(nameF)
+		msg, err := store.UpdateAccount(name, updatedName)
+		if err != nil {
 			return err
 		}
 
+		fmt.Println(msg)
 		return nil
 	},
 }